+++ src/components/date_picker/super_date_picker/super_date_picker.js
@@ -1,20 +1,20 @@
 
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
+import classNames from 'classnames';
 import { commonlyUsedRangeShape, recentlyUsedRangeShape } from './types';
 import { prettyDuration, showPrettyDuration } from './pretty_duration';
 import { prettyInterval } from './pretty_interval';
 
 import dateMath from '@elastic/datemath';
 
+import { EuiSuperUpdateButton } from './super_update_button';
 import { EuiQuickSelectPopover } from './quick_select_popover/quick_select_popover';
 import { EuiDatePopoverButton } from './date_popover/date_popover_button';
 
 import { EuiDatePickerRange } from '../date_picker_range';
 import { EuiFormControlLayout } from '../../form';
-import { EuiButton } from '../../button';
 import { EuiFlexGroup, EuiFlexItem } from '../../flex';
-import { EuiToolTip } from '../../tool_tip';
 
 export class EuiSuperDatePicker extends Component {
 
@@ -128,29 +128,6 @@ export class EuiSuperDatePicker extends Component {
     };
   }
 
-  componentWillUnmount() {
-    this._isMounted = false;
-  }
-
-  componentDidMount() {
-    this._isMounted = true;
-  }
-
-  setTootipRef = node => (this.tooltip = node);
-
-  showTooltip = () => {
-    if (!this._isMounted || !this.tooltip) {
-      return;
-    }
-    this.tooltip.showToolTip();
-  }
-  hideTooltip = () => {
-    if (!this._isMounted || !this.tooltip) {
-      return;
-    }
-    this.tooltip.hideToolTip();
-  }
-
   setTime = ({ start, end }) => {
     const startMoment = dateMath.parse(start);
     const endMoment = dateMath.parse(end, { roundUp: true });
@@ -174,11 +151,6 @@ export class EuiSuperDatePicker extends Component {
         this.props.onTimeChange({ start, end });
         return;
       }
-
-      this.showTooltip();
-      this.tooltipTimeout = setTimeout(() => {
-        this.hideTooltip();
-      }, 2000);
     }
   }
 
@@ -286,38 +258,16 @@ export class EuiSuperDatePicker extends Component {
       return;
     }
 
-    let buttonText = 'Refresh';
-    if (this.state.hasChanged || this.props.isLoading) {
-      buttonText = this.props.isLoading ? 'Updating' : 'Update';
-    }
-
-    let tooltipContent;
-    if (this.state.isInvalid) {
-      tooltipContent = 'Can\'t update, dates are invalid';
-    } else if (this.state.hasChanged && !this.props.isLoading) {
-      tooltipContent = 'Click to apply';
-    }
-
     return (
-      <EuiToolTip
-        ref={this.setTootipRef}
-        content={tooltipContent}
-        position="bottom"
-      >
-        <EuiButton
-          className="euiSuperDatePicker__updateButton"
-          color={this.state.hasChanged || this.props.isLoading ? 'secondary' : 'primary'}
-          fill
-          iconType={this.state.hasChanged || this.props.isLoading ? 'kqlFunction' : 'refresh'}
-          textProps={{ className: 'euiSuperDatePicker__updateButtonText' }}
-          disabled={this.state.isInvalid}
-          onClick={this.applyTime}
+      <EuiFlexItem grow={false}>
+        <EuiSuperUpdateButton
+          needsUpdate={this.state.hasChanged}
           isLoading={this.props.isLoading}
+          isDisabled={this.state.isInvalid}
+          onClick={this.applyTime}
           data-test-subj="superDatePickerApplyTimeButton"
-        >
-          {buttonText}
-        </EuiButton>
-      </EuiToolTip>
+        />
+      </EuiFlexItem>
     );
   }
 
@@ -336,10 +286,19 @@ export class EuiSuperDatePicker extends Component {
         isAutoRefreshOnly={this.props.isAutoRefreshOnly}
       />
     );
-    return (
-      <EuiFlexGroup gutterSize="s" responsive={false}>
 
-        <EuiFlexItem style={{ maxWidth: 480 }}>
+    const flexWrapperClasses = classNames(
+      'euiSuperDatePicker__flexWrapper',
+      {
+        'euiSuperDatePicker__flexWrapper--noUpdateButton': !this.props.showUpdateButton,
+        'euiSuperDatePicker__flexWrapper--isAutoRefreshOnly': this.props.isAutoRefreshOnly,
+      }
+    );
+
+    return (
+      <EuiFlexGroup gutterSize="s" responsive={false} className={flexWrapperClasses}>
+
+        <EuiFlexItem>
           <EuiFormControlLayout
             className="euiSuperDatePicker"
             prepend={quickSelect}
@@ -348,9 +307,7 @@ export class EuiSuperDatePicker extends Component {
           </EuiFormControlLayout>
         </EuiFlexItem>
 
-        <EuiFlexItem grow={false}>
-          {this.renderUpdateButton()}
-        </EuiFlexItem>
+        {this.renderUpdateButton()}
 
       </EuiFlexGroup>
     );